Darwin patches (from Jorge Luis Williams <jorgew@jlwg.net> ) only works for GLX apps -- it's still a ways to go before CGL support.

+++ crserverlib/get_components.py
@@ -134,7 +134,7 @@ for comp in comps:
 
 print """
 		default:
-			crError( "Unknown parameter name in lookupComponents: %d", pname );
+			crError( "Unknown parameter name in lookupComponents: %d", (int) pname );
 			break;
 	}
 	/* NOTREACHED */
